Fix the allversions problem.  For Plain.pm, the use of $files->getdir() in the recursive loop wasn't working, since at that $files->point no version was specified.
The fix is to iterate over the versions at the top level, using the
range variable from the config file to specify which versions are to
be indexed if the --allversions option (or none) is given.  Because
the value of "range" can be a closure, this should allow the CVS users
to dynamically discover the set of versions in a repository.