Fix bug 476695 - Java interfaces display as docs. Changes the type field in the indexes table to be a (langid, typeid) tuple, and adds another table to look up the (langid, typeid) -> string mapping. This means that each language module can manage its own string table and the right identifier type can easily be displayed.
This also means that each identifier is tagged with the language it occurs in,
which could be used for filtering displayed results when browsing.  This has
not been implemented.

This change is database incompatible - you will need to drop and recreate the
lxr database.  MySQL has been tested, Postgres has not though the changes have
been made.

Also add some limited support for Makefile in generic.conf.