Add automatic test framework and first tests to LXR. This depends on the Test::Unit suite from http://sf.net/projects/perlunit for the basic framework.  Currently an "AllTests" module is defined that should run all the tests currently available, as well as a ConfigTest module that makes a start on testing the LXR::Config.pm module.
Additional tests are planned and this should grow into a complete,
easy to run and maintain test suite that lets us catch regressions.

In future I plan to add a test to this suite for each bug reported, before
fixing it.  This (a) proves the bug exists and (b) makes it easy to see when
it has been fixed.  And it catches regressions when bugs come back.

To use this, you will need to:
1) Install Class::Inner from CPAN
2) Install Error from CPAN
3) Download the latest CVS version of Test::Unit from the URL above
4) Build and install Test::Unit from the CVS sources