Merge of the BK support from bk-dev-branch.
This merge closes the bk-dev-branch.

Adds support for BK repositories (somewhat incomplete).  As BK is no longer
free, this is untestable by the maintainer and will therefore not be maintained.
It is released here to allow others to take it forward if they want.

Also updates and add new testcases.