Allow for easy modification of line number width
In markupfile: pattern for @ltag adjusted so that line number may have any width with a minimum of 3 digits. All that is needed is to change $line above from
$line = "001"; to $line = "00001";
to get 5 digits. Interesting for files with more than 1000 lines to prevent the colums from being ragged (if tree contains many of these files).