CVS management: source file retrieval in side branches and automatic version enumeration
This fix involves many source files and subs, mainly in CVS.pm and genxref. The main task is to scan the CVS header to extract all tags and release numbers and to apply a new algorithm in parsecvs sub to reconstruct the target file.

genxref --allversions in CVS mode purges all DB tables. This necessitates a new method 'purgeall' in Files.pm (existent purge() operates on a single designated releaseid and we have no exhaustive list of releases for a CVS repository). This method has been tested for MySQL and copied as is in Oracle.pm and Postgres.pm. Testers welcome.

As a bonus, annotation display has been enhanced with background highlighting of latest addition to repository.