Common.pm: egg-and-hen problem
clean_release needs $pathname to check for an existing release and fix_paths needs and initialised $releaseid to determine the correct nature of $pathname (file or directory). To break the vicious circle, clean_release is computed first with a raw path from the URL and $pathname is subsequently canonised with a correct $releaseid.