Index.pm, Index/*.pm: Move common methods to Index.pm; add methods for incremental indexing
Common factor all DB-engine independent queries, keeping in the specific DB-backends only the idiosyncratic queries, notably those related with auto-increment since the syntax is specific to an engine.
Add methods to correctly erase definitions and usages in selected files of a single revision (to be used by genxref for incremental indexing).