Generic.pm, Java.pm, Perl.pm, Python.pm, Ruby.pm: fix include file processing
Fix for bug #231 (Java parser enters infinite loop when scanning "import static ..."): add optional static keyword in regexp for splitting include fragment.
While fixing this bug, a patch was needed on the "multi-directory path" handling because the trigger test was incorrect (testing for definedness on $link instead of an existing <a> HTML tag).
This fix is intended for 1.1.0->1.1.1 release, but the include feature will be rearchitectured for 1.2.0 because readability is poor.