source: UTF-8 truncation
This patch is an unperfect attempt to truncate author's name at a valid UTF-8 character boundary when displaying UTF-8. For other display encodings, truncation is made on a byte basis and may not be correct if an UTF-8 sequence is dumped into a non-UTF-8 stream. Reversely, truncation is erroneous for a non-UTF-8 string dumped into UTF-8 stream.

IMPORTANT: erroneous truncations may occur in diff since we do not take UTF-8 into account in sub htmljust().