genxref: make sure --version=x exists (bug #256 fix)
If an invalid version is passed through --version= option, genxref blindly
accepts what is returned by the $files-> functions without checking. The
functions results are "dummy" objects. Consequently, indexing empty files
is very fast and always succeeds, making the user believe the job is done.

Add a test to check for version existence in lxr.conf and, in the case of
plain files, that the source root directory is consistent with lxr.conf.
Abort if an error is detected.