Common.pm: change clean_release() algorithm
Sequentially explore the list of versions instead of creating a hash,
stuffing versions in it and then checking the candidate version.
The rationale is the number of versions in usual cases is rather small and
sequentially exploring the list avoids the penalty of creating and
accessing the hash.
The old code is kept as a commented out block for "pathological" cases
such as Linus's tree.