genxref, Index.pm, Index/*: open DB read-only whenever possible
When browsing a source tree, no modification is done on the DB. It is then
pointless to update the "unique counters". It is even better not to do it
when DB is stored in standard files (e.g. SQLite) as access rights may not
allow it: web server runs under its own user-id which is considered "others"
from LXR administrator point of view. This allows others' permissions to
be set to read-only, preserving DB integrity.