ident, search, templates/html/*ident|search*: improve landing page display
When ident or search are directly entered (i.e. no search argument provided
yet), there can be no result. It is then simpler to provide HTML attribute
"hidden" so that templates can have the information and accordingly disable
(or hide) the result area.

This results in a new marker $noquery, expanding to string "hidden" if no
search argument is passed to the script. Typical usage is to embed the
result area inside a <div $noquery> ... </div>.