genxref: protect DB against integrity threat
Since unique numbering of files, symbols and types is now managed uniformly
across DB engines through custom tables and procedures, saving the numbers
is required to avoid duplicate key errors on subsequent execution.

The easiest way is to commit the tables whenever a value changes, but this
also commits the rest of the cached transactions, resulting in a dramatic
loss of performance.

To keep the cache performance advantage, any abort signal must be caught so
that the current counter values can be sent to the DB.

A __DIE__ handler is enabled in genxref critical sections to call $index->
finalcleanup() to do it.