Lang/Generic.pm: bug fix for ctags call with shell special chars
ctags command line is parsed by the shell. Consequently shell special
characters in filenames must be escaped so that they are not interpreted.

Note: call to diff in script diff does not incur the same problem because
diff utility is invoked through exec() which does not pass its arguments
to the shell.