templates/initdb/initdb-m-template.sql: work around TRUNCATE TABLE bug
Although TRUNCATE TABLE should be very fast, a bug in MySQL/MariaBD causes
it to become extremely slow on huge tables. It is recommended to replace
the statement by a RENAME TABEL, create TABLE LIKE and DROP TABLE.

Also, creation of trigger/procedure/function is made consistent: every
operation is preceded by a DROP xxx IF EXISTS so that creation is not
rejected if the object already exists and DB definition state is reliable.