Local.pm: fix for bad default description processing
Default description is preloaded with HTML entity reference &nbsp; then
a starting fragment of the file is added to this default content.
Description extractions progressively removes "meaningless" runs of
characters, among which special characters like &, suppressing the entity
reference significance and leaving a weird description.

Fix: do not preload a default but check that returned value is at least
&nbsp; if the final description is empty.