Common.pm: bug fix for some blank screens (did not send HTTP headers)
Multiple errors occurring before HTTP is fully initialised may result in
no headers being sent to the server. This is the case if a warn precedes
a die statement. It was thought that this sequence was unlikely because
only die statements are used early and warning() intercepting function did
not check HTTP initialisation. This assumption neglected library routnies
which potentially emit warn()!

Fix: test in warning() if HTTP is initialised. If not, call httpminimal()
to issue a minimal set of headers before printing anything.