Index/Mysql.pm, templates/init/initdb-m-template.sql: bug fix in DB purge
Commit 371a29 fixed the MySQL purge performance issue but introduced a bug
caused by an undocumented effect of DROP (or RENAME?) TABLE: all TRIGGERS
associated with the table are also erased. Triggers must be re-created.

Fix is not satisfactory because CREATE TRIGGER cannot be done from within
a procedure. Trigger creation must then be done in two locations with the
risk of forgetting to update one.
A reminder has been added in the template to warn about it.