Index.pm, Inex/*: full DB purge through recreation
When a DB is dedicated to a tree (no other tree hosted), it is faster to
drop the DB than deleting the records or tables. Thr DB is then recreated
using the shell script saved in custom.d/db-scripts.d.

We must then disconnect from the DB before launching the script, then
reconnect to it and rebuilds the prepared transactions. Since the same
DB object is reused, detecting overriden transactions can no longer be
done with exists() function (which would return true for the stale
transactions). This requires some rewrite of the initialisation methods.

Remaining issue with PostgreSQL: post_processing() requires password
passing to perform DB optimisation. Despite the clues given in the manual,
password passing was unsuccessful, which results in an unacceptable prompt
which stalls unattended operation.
As a temporary measure in genxref (separate commit from this one), option
-U (=unattended) option was introduced. For PostgreSQL DBs, it is
equivalent to --novacuum to skip the step. It does nothing on other DBs.