genxref: minimize downtime on glimpse indexation
Glimpseindex job may be very long on huge trees and users may still be
querying the database. It is unacceptable to make them wait for more than
a few seconds. Consequently, it is preferable to index "offline" and then
switch the indexes. Even if this results in inaccurate results during
indexing, it is much better then shutting down the site.

Create 'glimpseindexopts' in lxr.conf to offer flexibility in indexing