templates/initdb/initdb-*.sql: undefined variables and *_times table
Protect shell if expressions against NO_DB and NO_USER undefined variables
which would cause syntax error when used in expressions. Also write
expressions in a more portable dialect.

Always drop *_times table when reconstructing DB: the tables will be erased
by genxref anywa. Dropping them allows to udate the schema systematically.
Note that contents is not related to the tree itself, so that keeping the
statistics data is not a great issue.