base CGIs, Common.pm, Template.pm: fix missing template handling
Simplify surrogate template management: if requested template not found,
print a warning and return the surrogate argument.
In case the missing template is one of the "heads", <html> and <body> are
missing. They are provided by warning() or fatal(). Therefore, $wwwdebug
should no longer be set to zero (except a very specific case with thttpd)
otherwise the resulting HTML stream will be buggy.

gettemplate() interface has been changed because there is no use for a
"prefix" to the template: it can't be sent before warn or die and after
these we have the "suffix" into which the "prefix" can be merged.

Make sure $HTMLheadOK is really undefined at start so that detection is
correct.