Common.pm, diff, ident, search, source: change HTTP header generation and      index state detection
For "dynamic" scripts like perf, it is necessary to send an HTTP
Last-Modified header set to the currecnt time so that the browser does not
use a stale copy. Add the script name as an argument to std_http_headers().

Improve index state assessment by not retrieving text or usage records when
this is not relevant: the perf case should exit as soon as it got (or not)
the global progress record.