genxref: experimental multi-process code
Parallel free-text and references indexing was explored. Results determined
that LXR indexing is I/O-bound and no gain can be expected with such a
strategy (the sub-tasks compete for the I/O subsystem and disrupt their
flows leading to a slightly longer total elapsed time).

Commented code is kept for reference.