** New feature (cont.) ** Free-text search with VCS (add Subversion)
Change name of method to exporttree() because 'checkout' is not
only an export in Git but also implies to modify internal state.

Fix GIT.pm so that it is really "transparent", not affecting
internal state or local modifications.

Add Subversion method and fix genxref to cope with the virtual
file system paradigm.