Lang/Perl.pm, */Python.pm: fix subtle endless loop bug
Under some weird circumstances, processinclude() may cause the web
server to be locked in a seemingly endless loop. Superficial
investigation incriminates path rewriting when only a directory is
specified. Not adding a file extension seems to cure the problem.

However, keep an eye on it as the explanation is not clear.