genxref: fix duplication of glimpse databases
When in incremental mode (no --reindexall option), glimpse databases
are first copied to an "offline" location so that glimpse can proceed
without disturbing running "online" service. All DBs coming from a
"version directory" are now a subdirectory in the temporary one.

After glimpse indexing, newer DBs are substituted for old ones. This
was done with a mere 'cp' on the temporary directory which adds one
directory level in the hierarchy.

The correct sequence is 'mv' on the DB subdirectory (one level deep in
the temporary) followed by a 'rmdir' on the temporary.