Lang.pm: fix for bug 276, endless loop on multiple path separator
When an "include" file path contains several separators in a row,
sub _linkincludedirs enters an endless loop because it is unable
to remove extra separators without name (it expects path fragments
to be separated by exactly one delimiter).

A quick fix could be to remove a fragment preceded by any number of
separators but this leads to non-optimal HTML rendering of the
path. The final fix carefully checks the separators so that the
"active" parts of the statement contain only fragment names.

Note: coding several separators is bad practice; however OSes don't
consider this as an error and interpret correctly the path.